Bug fixes and minor performance improvements
Sensor and page ID type fixes that caused many issues. Usage of hash maps to increase look-up times on sensors and pages for many operations e.g. setting sensor data.

+++ src/com/bennero/server/message/PageSetupMessage.java
@@ -75,8 +75,8 @@ public class PageSetupMessage {
         Color titleColor = Color.rgb(titleColourR, titleColourG, titleColourB);
         Color subtitleColor = Color.rgb(subtitleColourR, subtitleColourG, subtitleColourB);
 
-        PageData pageData = new PageData(pageId, backgroundColour, titleColor, subtitleColor, pageRows, pageColumns,
-                nextPageId, pageTransitionType, pageTransitionTime, pageDurationMs, title,
+        PageData pageData = new PageData((byte)pageId, backgroundColour, titleColor, subtitleColor, pageRows,
+                pageColumns, (byte)nextPageId, pageTransitionType, pageTransitionTime, pageDurationMs, title,
                 titleEnabled == 0x01, titleAlignment, subtitle, subtitleEnabled == 0x01,
                 subtitleAlignment);
